Deduplicate ID header transform
